face_recognition_app/requirements.txt
fastapi
uvicorn
python-multipart
insightface
onnxruntime
qdrant-client
numpy
opencv-python-headless
pydantic
requests
pytest
httpx
pillow-heif
Pillow